requirements.txt
distlib==0.3.8
filelock==3.13.1
lxml==5.1.0
platformdirs==4.1.0
PyMuPDF==1.23.17
PyMuPDFb==1.23.9
python-docx==1.1.0
typing_extensions==4.9.0
virtualenv==20.25.0
